+++ plugins/fr.cea.nabla.ui/examples/NablaExamples/src-gen-kokkos/glace2d/Glace2d.cc
@@ -150,7 +150,7 @@ private:
 			int jId(jCells);
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rMinus1Id(nodesOfCellJ[(rNodesOfCellJ-1+nbNodesOfCell)%nbNodesOfCell]);
 					int rPlus1Id(nodesOfCellJ[(rNodesOfCellJ+1+nbNodesOfCell)%nbNodesOfCell]);
@@ -189,7 +189,7 @@ private:
 			int jId(jCells);
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rMinus1Id(nodesOfCellJ[(rNodesOfCellJ-1+nbNodesOfCell)%nbNodesOfCell]);
 					int rPlus1Id(nodesOfCellJ[(rNodesOfCellJ+1+nbNodesOfCell)%nbNodesOfCell]);
@@ -225,7 +225,7 @@ private:
 			int jId(jCells);
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					l(jCells,rNodesOfCellJ) = MathFunctions::norm(C(jCells,rNodesOfCellJ));
 				}
@@ -247,7 +247,7 @@ private:
 			double reduction5 = 0.0;
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rId(nodesOfCellJ[rNodesOfCellJ]);
 					int rNodes(rId);
@@ -274,7 +274,7 @@ private:
 			RealArray1D<2> reduction0 = {{0.0, 0.0}};
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rId(nodesOfCellJ[rNodesOfCellJ]);
 					int rNodes(rId);
@@ -295,7 +295,7 @@ private:
 			double reduction1 = 0.0;
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rId(nodesOfCellJ[rNodesOfCellJ]);
 					int rNodes(rId);
@@ -442,7 +442,7 @@ private:
 			int jId(jCells);
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					Ajr(jCells,rNodesOfCellJ) = ArrayOperations::multiply(((rho(jCells) * c(jCells)) / l(jCells,rNodesOfCellJ)), tensProduct(C(jCells,rNodesOfCellJ), C(jCells,rNodesOfCellJ)));
 				}
@@ -464,7 +464,7 @@ private:
 			double reduction2 = 0.0;
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					reduction2 = reduction2 + (l(jCells,rNodesOfCellJ));
 				}
@@ -487,7 +487,7 @@ private:
 			RealArray2D<2,2> reduction3 = {{{0.0, 0.0}, {0.0, 0.0}}};
 			{
 				auto cellsOfNodeR(mesh->getCellsOfNode(rId));
-				for (int jCellsOfNodeR=0; jCellsOfNodeR<cellsOfNodeR.size(); jCellsOfNodeR++)
+				for (size_t jCellsOfNodeR=0; jCellsOfNodeR<cellsOfNodeR.size(); jCellsOfNodeR++)
 				{
 					int jId(cellsOfNodeR[jCellsOfNodeR]);
 					int jCells(jId);
@@ -513,7 +513,7 @@ private:
 			RealArray1D<2> reduction4 = {{0.0, 0.0}};
 			{
 				auto cellsOfNodeR(mesh->getCellsOfNode(rId));
-				for (int jCellsOfNodeR=0; jCellsOfNodeR<cellsOfNodeR.size(); jCellsOfNodeR++)
+				for (size_t jCellsOfNodeR=0; jCellsOfNodeR<cellsOfNodeR.size(); jCellsOfNodeR++)
 				{
 					int jId(cellsOfNodeR[jCellsOfNodeR]);
 					int jCells(jId);
@@ -610,7 +610,7 @@ private:
 			RealArray1D<2> nY = {{0.0, 1.0}};
 			{
 				auto nodesOfFaceF(mesh->getNodesOfFace(fId));
-				for (int rNodesOfFaceF=0; rNodesOfFaceF<nodesOfFaceF.size(); rNodesOfFaceF++)
+				for (size_t rNodesOfFaceF=0; rNodesOfFaceF<nodesOfFaceF.size(); rNodesOfFaceF++)
 				{
 					int rId(nodesOfFaceF[rNodesOfFaceF]);
 					int rNodes(rId);
@@ -664,7 +664,7 @@ private:
 			int jId(jCells);
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rId(nodesOfCellJ[rNodesOfCellJ]);
 					int rNodes(rId);
@@ -702,7 +702,7 @@ private:
 			double reduction7 = 0.0;
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					int rId(nodesOfCellJ[rNodesOfCellJ]);
 					int rNodes(rId);
@@ -727,7 +727,7 @@ private:
 			RealArray1D<2> reduction6 = {{0.0, 0.0}};
 			{
 				auto nodesOfCellJ(mesh->getNodesOfCell(jId));
-				for (int rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
+				for (size_t rNodesOfCellJ=0; rNodesOfCellJ<nodesOfCellJ.size(); rNodesOfCellJ++)
 				{
 					reduction6 = ArrayOperations::plus(reduction6, (F(jCells,rNodesOfCellJ)));
 				}
